Move subscriber into its own file.
