+++ spec/lib/subscriber_spec.rb
